requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-dotenv>=1.2.1
Pillow>=10.0.0
openai>=1.0.0
python-multipart>=0.0.6
